requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.2
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.0.1
pytest-django==4.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
whitenoise==6.6.0
